Add spinoso-exception crate
`spinoso-exception` extracts the structs that define the base
`Exception` types in Ruby Core into a separate crate. These types are
unified with a `RubyException` trait that offers `message` and `name`
APIs.

The exception sources are largely autogenerated, but avoiding macros
allows the docs and types to be evolved independently, for example to
add a `name` field to `NameError`.

One not great part of this crate is the duplication of `literal.rs` from
`spinoso-symbol`. A crate that collects common parsing utilities will be
useful.

This commit fixes a bug where the `Fatal` exception name was registered
as `Fatal` instead of a lowercase `fatal` as it is defined in MRI.

A followup commit will replace the types defined in
`artichoke_backend::extn::core::exception`.