Remove trait objects from the artichoke-backend virtual filesystem
This commit removes the `Filesystem` trait from `artichoke_backend::fs`.
This trait is a holdover from long ago when Artichoke depended on the
`filesystem` crate.

This significant refactor of the `fs` module follows the same approach
I've been using in the Spinoso crates: make multiple backends that are
source compatible so they can be swapped out with Cargo feature flags.

This change introduces three `fs` backends:

- `Memory`: an in-memory backend that stores file contents in a
  `HashMap` and uses a `HashSet` to track `$LOADED_FEATURES`. `Memory`
  filesystems support storing extension hooks, which are used to
  initialize gems in Artichoke.
- `Native`: a host filesystem adapter that stores and reads file
  contents from local disk and uses a `HashSet` to track
  `$LOADED_FEATURES`.
- `Hybrid`: a composite `fs` backend that mounts a `Memory` instance at
  a prefix and forwards all operations to it that match this prefix. All
  other filesystem operations target the host filesystem.

This change also stores paths inside the `fs` module as `BString`. This
means paths on Windows must be valid UTF-8 so they can be normalized
(converting `\` to `/`). I expect this to fix a good many `should_panic`
tests once this branch gets pushed to CI.