Add scolapasta-hex crate
Start to split out reusable utility functions into their own crates.

Previously, the approach was to copy and paste source files between
crates, like `literal.rs` for String escapes in `spinoso-exception` and
`spinoso-symbol` (and eventually `spinoso-string`).

With the push to add `const` annotations to as many functions as
possible, having duplicate copies of sources is error prone.

This change extracts the `hex` module from `spinoso-securerandom` into
its own crate so it can be resused in `spinoso-random` once the `Random`
core class is extracted from `artichoke-backend`.

This change enables the addition of cargo features. This also means that
`spinoso-securerandom` does not need to export the `hex` module to keep
some functions reachable (like `hex::write_into`) that it does not use.

The `EscapedByte` iterator was modified to use a `Chars` iterator
internally, which simplifies the iterator implementation significantly.

`<EscapedByte as From<u8>>::from` was lifted into
`EscapedByte::hex_escape` and made const. `hex_escape` returns a
`&'static str`.

This new configuration will make the `EscapedByte` iterator easier to
transition to a moved iterator of `[char; 2]` once this construction is
stable.