Build mrblib C sources
The Rust 1.46.0 upgrade on Windows somehow uses a stricter linker than
before. This has been a bug for a long time, but the functions involved
with the missing 'mrblib_irep' symbol have never been used.

`mrblib_irep` is defined as an `extern const` array. This commit
provides a valid C89 definition for this symbol in
`mruby-sys/src/mruby-sys/extn.c` that can be resolved at linkage.

The linker error might come from the recent update to the GitHub Actions
VM to Version: 20200820.1.