Implement Random in artichoke-backend with spinoso-random
Implement the `Random` core class in `artichoke-backend` extn using
`spinoso-random`.

This change removes the `rand_pcg` dependency. Artichoke now uses the
same reproducible Mersenne Twister PRNG asd MRI.

`artichoke-core` is bumped to 0.5.0 with breaking changes to the `Prng`
trait. The trait is reworked to have an associated type parameter for
the PRNG and only two methods for accessing the inner PRNG: `Prng::prng`
and `Prng::prng_mut`. This significantly simplifies the implementation
of the trait and `extn::core::random` by removing indirection. This
change will also be useful once `spinoso-random` adds support for range
`Max` constraints.

`artichoke-backend` no longer directly depends on `rand`.
`spinoso-random` uses `rand` and `rand_core` as an implementation detail
and exposes a full API to `artichoke-backend` that does not require
importing an `rand` traits.

This commit adds `AsRef<T>` and `AsMut<T>` impls to `UnboxedValueGuard`
for ergonomics when dealing with references after unboxing. This is used
to access the `PartialEq` impl on the `Rng` type in `extn::core::random`.

Instead of representing the interpreter global PRNG with a trait object,
this commit uses an `Rng` enum. An empty `Rng::Global` variant forwards
to the interpreter state. This dispatch happens in the trampoline and is
transparent to the PRNG implementation.

Seeds and constructors were reworked to fit into `spinoso-random`'s
recommended 128-bit seed, while recognizing the limitations of
Artichoke's lack of `Bigint` support.

Constructing the state is now fallible, recognizing that reading a
random seed from the platform may fail due to insufficient entropy.

In `spinoso-random`, `rand_core::Rng` impls were split out into a
separate `rand_core` feature. Added `impl Default for Random` that tries
to generate a random seed from the platform source of randomness and
falls back to the reference MT19937 seed on failure.