Remove generic bound in artichoke_backend::interpreter_with_config
Remove the generic bound on `ReleaseMetadata` in
`artichoke_backend::interpreter_with_config` and
`artichoke_backend::extn::init`. Replace it with taking the
concrete `ReleaseMetadata` type defined in `artichoke-backend`.

Implement all of the `core::ReleaseMetadata` trait methods as `const fn`
associated methods on `ReleaseMetadata`. Combined with taking the
concrete type, this should allow const propagation to fill in the Ruby
constants.

Followup to #703.