Remove dependency on uuid crate
Remove the dependency on the `uuid` crate and implement the UUID V4
generation algorithm directly with `rand`.

The UUID V4 generation algorithm is ~25 lines and is not worth the cost
of an extra dependency.

Pulling this code into `spinoso-securerandom` drops another dependeny
that indirectly depends on `rand` and `getrandom`. Both of these crates
have pending semver incompatible updates. Removing transitive deps on
these crates will make orchestrating the upgrade easier.

The code change in `artichoke-backend` to consume this new change, which
properly exposes the fallibility of random bytes generation, is trivial
(the addition of a single `?` operator to propagate the error).

This change adds property tests to ensure the generated UUIDs serialize
to the correct format with the correct version and generation tags.

This change adds a new `escape_byte` free function to `scolapasta-hex`
to expose the inner hex escaping APIs for single bytes.