Enable clippy::needless_borrow lint
Enable `clippy:needless_borrow` lint to check for unnecessary references
that are immediately dereferenced by the compiler.

I first asked about this in rust-lang/rust-clippy#4271 but never got
around to enabling it.

This lint is in the nursery lint group, so explicitly activate it in all
`lib.rs` sources for all workspace crates.

This lint caught a few needless borrows in `artichoke-backend`'s
interpreter infrastructure and `artichoke`'s Ruby CLI.