Add raw parts APIs to spinoso_array::Array
Add `Array::from_raw_parts` and `Array::into_raw_parts` APIs to the
`Vec`-based `Array` type in `spinoso-array`.

These APIs may be used to turn an `Array` into a raw pointer that can
later be reconstructed to an `Array`. These APIs are highly unsafe and
calling these APIs moves ownership of the underlying allocation.

These APIs are most useful in FFI contexts and are being introduced to
prepare for fixing GH-922.

Both `from_raw_parts` and `into_raw_parts` are marked must use to force
callers to acknowledge drop/leak intent with a wildcard binding.

The corresponding raw parts APIs cannot be added to the `SmallArray`
type because its internal `SmallVec` can only soundly be decomposed into
a raw parts triple when the storage is spilled to a `Vec`. This would
force users of `SmallArray` to forcibly spill the `SmallVec` to the
heap, which eliminates much of the usefulness of the `SmallVec`
container. Additionally, `SmallVec` does not expose an API to force a
spill to the heap.

The addition of these APIs bumps the `spinoso-array` crate version to
0.4.0.