Amortize double pointer deref on Artichoke State
The inner Rust `State` is stored as a `Option<Box<State>>` property on the
`Artichoke` interpreter. When `artichoke-backend` implements interpreter
traits from `artichoke-core`, it is common to access the inner `State`
using `Option::as_ref` and `Option::as_mut` at the beginning of a method
to unwrap the `State` or error out. These APIs return `&Box<State>` and
`&mut Box<State>` (aka pointers to a pointer), which results in a double
indirection.

Replace all calls to unwrap the `State` with `Option::as_deref` and
`Option::as_deref_mut` to retrieve a `&State` and `&mut State`.
Implementations of `artichoke-core` traits often unwrap the `State` once
before performing several operations on it. This change results in
following the double pointer once via `Deref` to amortize its cost over
the entire function call.