Remove GC disable from BoxUnboxVmValue impl for HeapAllocatedData
GH-707 introduced a hack to the `BoxUnboxVmValue::alloc_value`
implementation for `HeapAllocatedData` to disable GC before allocating
an `MRB_TT_DATA` object.

This hack had to be introduced because a borrow on the interpreter must
be held to access the class's `mrb_data_type` which precluded the
implementation from using `Artichoke::with_ffi_boundary` to repack the
state. The `State` must have been repacked because `Array` was
implemented as `MRB_TT_DATA` and required the `State` to free Artichoke
`Array`s.

As of this commit, both of these invariants are no longer true:

- `mrb_data_type`s are leaked to given them static lifetime which
  enables wrapping the call to `mrb_data_object_alloc` in
  `Artichoke::with_ffi_boundary`.
- Artichoke `Array` is implemented with `MRB_TT_ARRAY` and can free the
  underlying `Vec` directly via `Vec::from_raw_parts` instead of
  resolving via the `State`.

This commit removes the GC disable and explanation of its purpose.