Remove Clone derive on class and module specs
Calling `spec.clone()` and using this cloned spec to initialize objects
with `sys::mrb_data_object_alloc` is unsafe if the cloned spec is
dropped before the interpreter is closed because a `dfree` function
pointer associated with an `mrb_data_type` may be deallocated before
the object is collected.

This commit reworks the `EnclosingRubyScope` to store the components of
the class and module specs it requires, which does not include the
`mrb_data_type`.

This refactor was made easier by #793 which moved the
`EnclosingRubyScope` `Box` to the `EnclosingRubyScope` fields that
require it instead of boxing in the specs.