Reimplement spinoso_symbol::Inspect to use raw bstr decode UTF-8 APIs
Reimplement `spinoso_symbol::Inspect` to use `bstr::decode_utf8` and
`bstr::decode_last_utf8` instead of `bstr::CharIndices`. This reduces
the size of `spinoso_symbol::Inspect` from 136 bytes to 120 bytes.

Despite the APIs used being lower level, the code size doesn't increase
and the match arms are a bit more obvious in that they don't need
comments explaining what they do and why they are implemented the way
they are. The match arms also remove a special branch for the Unicode
replacement character which is likely to be fairly cold.