Simplify implementation of String#ord
Reimplement `String#ord` in the artichoke-backend string trampoline to
use `bstr::decode_utf8` instead of `bstr::ByteSlice::char_indices`.

This fixes a bug where `String`s that began with the Unicode replacement
character did not return 65,533.

This change also properly raises `ArgumentError` with message "invalid byte
sequence in UTF-8" when the `String` begins with an invalid UTF-8 byte
sequence. This change assumes that all Artichoke `String`s implicitly have
UTF-8 encoding. Work is underway on `spinoso-string` which explicitly
tags `String`s with either UTF-8, ASCII, or binary encoding.

This commit updates the link to the `char` converter source in
rust-lang/rust to 1.48.0, which accounts for the source directory layout
changes that took place since the link was generated with the 1.41.0
tag.