Convert spec-runner configuration file from YAML to TOML
This commit makes a breaking change to the spec-runner by changing the
configuration format from YAML to TOML. This commit bumps the version
number of the `spec-runner` to 0.3.0.

The YAML structure for declaring tagged specs is not the best. This
commit converts the structure from:

`Map<Vec<NamedSuite>>`

to:

`Map<Family, Map<Suite, Specs>>`

using TOML as the configuration format.

This commit removes `serde_yaml` and its dependencies and replaces it
with the `toml` crate, which has zero dependencies.

The config format could be translated to TOML and the model reused as
is, but the resulting TOML would have lots of e.g. `[[core]]` nested
table blocks which are a bit ugly. This new configuration format ensures
that each API suite is only declared once.

The `model` module is reimplemented for the new config format and the
model and spec setup code is refactored a bit to improve readability.