Add a CI job to compile Artichoke with its minimum dependency versions
Ensure that Artichoke and the spec-runner compile with the lowest valid
dependency versions based on the constraints in `Cargo.toml`.

Dependency constraints are a contract that declares any satisfying
version is valid for building the crate. If this is not true, the
dependency specifiers are buggy.

This PR updates the minimum dependency constraints for several crates:

- `bstr` in `artichoke-backend` moved from `0.2` to `0.2, >= 0.2.2` to
  ensure `impl Default for &'a Bstr` and `impl Default for BString`.
- `intaglio` in `artichoke-backend` moved from `1` to `1, >= 1.1.1` to
  remove the dependency on `bstr`, which improperly specified its
  minimum version as `0.2` when it relied on `impl Default for &'a BStr`
  which was introduced in `0.2.2`.
- `log` in `artichoke-backend` moved from `0.4` to `0.4, >= 0.4.5` to
  pull in a fix for an inner macro resolution compile error.
- `serde` in `spec-runner` moved from `1.0` to `1.0, >= 1.0.7` to work
  around a bug in minimal vesions of `toml` which improperly declare the
  minimum version of `serde`.

This commit adds a non-blocking CI job which tests both `artichoke` and
`spec-runner` compile and test compile with their minimal versions by
forcibly generating a minimal-versions lockfile using nightly `cargo`.