Ensure a newline is written after the calling inspect on the last value
In v7.0.0 and later, `rustyline` overwrites lines in the terminal that
are not newline-terminated (not flushed?). This results in an `airb`
REPL that doesn't display the `#inspect` value of the last expression.

This commit fixes this bug by writing a `"\n"` to the output stream
after writing the `#inspect` value of the last expression.

Prior to this commit:

```console
$ cargo run -q --bin airb
artichoke 0.1.0-pre.0 (2020-12-24 revision 3889) [x86_64-apple-darwin]
[rustc 1.48.0 (7eac88abb 2020-11-16)]
>>> "a"
>>>
```

After this commit:

```console
$ cargo run -q --bin airb
artichoke 0.1.0-pre.0 (2020-12-24 revision 3890) [x86_64-apple-darwin]
[rustc 1.48.0 (7eac88abb 2020-11-16)]
>>> "a"
=> "a"
>>>
```