Re-enable clippy::or_fun_call
Disabling `clippy::or_fun_call` lint for calls to const functions was
done with an incorrect understanding that const fns with non arguments
would be guaranteed to be const evaluated. This is not the case. See
rust-lang/rust-clippy#6077.

The revert of this behavior in `clippy::or_fun_call` will land in Rust
1.49.0 due out in a week. This commit removes the pragmas to disable the
lint and modifies all callsites to use the lazy `ok_or_else` variant.

This commit reverts a change made in #794 and 9a8c14c9146.