Test for memory leaks with LeakSanitizer in CI
This commit removes the bespoke memory leak checker in
`artichoke-backend` integration tests implemented with `libc::getrusage`.
This leak checker was introduced in 468169d22233b941df12c92b69955ffba123d6cc.

This checker was inaccurate and only enabled for linux targets.

This commit removes the `leak` module and adds a CI job to run all of
the tests in the Artichoke workspace with LeakSanitizer on nightly Rust.

This commit refactors some integration tests to break apart individual
test cases now that the tests are no longer constrained to spawn a
single thread for leak checking accuracy.

Followup to #933 and #937.