Use with_message constructor from spinoso-exception for &'static str msg
For messages that are `&'static str`, use the `<Exception>::with_message`
constructor.

This constructor expresses the intent and type bound that the exception
message is:

- A `&'static str`.
- A UTF-8 string.

`with_message` is also const, whereas the `From` impl is not, which at
least gives `rustc` a chance to do const propagation.

Followup to #958 which introduced this API.