Remove conditional application of no_std crate-level attribute
The preferred mechanism for conditionally depending on `std` is to
always declare `no_std` and conditionally link to `std` with an
`extern crate std;` declaration.

This makes the `lib.rs` prelude a bit simpler and makes it easier to
depend on `std` in tests. Additionally, items from `std` must always be
imported regardless of features, which might have masked bugs.

Crates and tests might not compile as of this commit.