Restructure scolapasta-hex tests so they compile with any feature set
Before this commit, the `scolapasta-hex` test suite only compiled and
ran with the `std` feature enabled.

This commit breaks apart the test suite into several modules `alloc` and
`std` that correspond to the features that must be enabled to execute
the tests.

Additionally, individual functions are broken out into their own test
functions on the same set of test vectors.