Add caching to GitHub Actions workflows
This commit adds caching for the cargo registry, git cache, and target
directory of pre-built dependencies for the root workspace, fuzz
workspace, and spec-runner workspace.

These changes are made for the CI and rustdoc workflows.

These changes should speed up CI iterations significantly. In the
playground repo where these actions steps are cribbed from, this led for
a 50% reduction in build times.