Add Rust compiler artifact caching to GitHub Actions CI
Add caching using the `Swatinem/rust-cache@v1` GitHub Action for all CI
jobs that use a stable compiler.

`rustdoc`, `fuzz`, and `leak-san` jobs use a nightly compiler which is
updated daily and will not get much reuse out of the `rust-cache`-generated
cache keys which include the rustc compiler hash on nightlies.

This GitHub Action was introduced successfully in the playground in
artichoke/playground#299.

This commit is a reimplementation of the failed PR #1016 that works
around actions/cache#403.

The build matrix has a `v3` cache key to bust the cache while rolling
out the cache on windows. See Swatinem/rust-cache#7.