Fix interpreter memory leaks in REPL and Ruby frontends
Neither `airb` nor `artichoke` binary frontends closed the `Artichoke`
interpreter, which leaks the interpreter and its heap. This does not
have many consequences since the program is exiting anyway, but it does
prevent destructors for Ruby values from being called.

This fix restructures the `ruby` and `repl` modules so they create an
interpreter at the top level and pass a `&mut Artichoke` to the
entrypoint implementation. This allows all short circuit `Err` returns
to be caught so the interpreter can be closed before passing the
`Result` back to `main`.