Add the beginnings of the spinoso-string crate
This commit adds a WIP spinoso-string crate, which will be used to
implement the `String` Ruby Core class in Artichoke.

Spinoso `String` is modeled as a `Vec<u8>` with an associated
`Encoding`. The `Encoding` is conventional only, but does influence
several read-oriented APIs, like `String#ord` and `String#chr` as well
as new constructors like `String#center`.

Because a Spinoso `String` is meant to be packed into an mruby `RArray`,
it can be converted to raw pointer. `Encoding` can be serialize to a
`u8` bit flag.

Spinoso `String` intentionally only supports three Ruby encodings:

- UTF-8
- ASCII
- Binary / ASCII-8BIT

For UTF-8 support, `spinoso-string` makes use of `bstr`.