Implement Array#concat in Rust
This commit implements `Array#concat` in Rust and works around previous
issues regarding concatenating to `self` by allocating a new `Array` and
concatenating into it.

This commit moves the core `Array#concat` implementation to the
`array::trampoline`, which removes a method which requires the
interpreter from the `Array` data structure.