Make spinoso-regexp::Debug an Iterator
Rather than a struct with an iter method that returns an Iterator.

This change will make `debug()` functions on Regexp backends behave as
is typical for functions that return structs named the same as the
function: the returned struct is an iterator. Previously, `Debug` had no
useful methods on it other than `iter()` and an `IntoIterator`
implementation. This change removes the indirection.