Refactor artichoke-backend integration tests to reduce compile/run time
This commit refactors the integration tests for `artichoke-backend` into
a single crate. On my machine this reduces compilation time for
`cargo test --workspace --no-run` by 3% and reduces the runtime of
testing the entire workspace following a build with `cargo test --workspace`
by 12%. These two improvements total about 9 seconds in saved build and
run time.

https://www.reddit.com/r/rust/comments/lto0qa/blog_post_delete_cargo_integration_tests/
https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html

This refactor takes an opportunity to unify crate-level attributes,
remove `deny` attributes, and update the comments associated with the
tests to refer to APIs and structs with their current names.