Add a RUBYLIB vfs Ruby source loader
Add a new vfs loader that searches the `RUBYLIB` environment variable
for sources in `require` and `relative_require`.

The `RUBYLIB` loader is only activated if the `RUBYLIB` environment file
is set on interpreter boot. The `RUBYLIB` loader has the highest
priority of all vfs loaders.

This patch enables the Artichoke CLI to execute the spec suite without
the `spec-runner` crate and the embedded sources!

```console
$ RUBYLIB=spec-runner/vendor/mspec/lib:spec-runner/src ./target/release/artichoke -e 'require "spec_runner"' -e 'run_specs []'
artichoke 0.1.0-pre.0 (2021-03-20 revision 4242) [x86_64-unknown-linux-gnu]

Passed 0, skipped 0, not implemented 0, failed 0 specs.
```

The `RUBYLIB` loader is gated behind an on-by-default cargo feature. To
disable the `RUBYLIB` loader, build without the `rubylib-source-loader`
feature.