Refactor alternate require path resolution into LoadSources impl
The functions that implement `require` and `require_relative` in `extn`
manage retries by adding a missing `.rb` extension to the file path.

The `LoadSources` trait that is implemented by `Artichoke` in
`artichoke-backend` is meant to fully encapsulate the loading of Ruby
sources from the underlying virtual filesystem.

As I have attempted to add support for extending Artichoke's
`$LOAD_PATH` via the `RUBYLIB` environment variable, I ran into issues
with the `Hybrid` vfs. The hybrid backend does not allow its component
backends to absolutize paths relative to their own current working
directory.

This deficiency shows up in the `require` and `require_relative` impls
in that they manually absolutized paths relative to the `Memory`
`RUBY_LOAD_PATH` constant. With the future addition of a `Rubylib` vfs
backend, it is no longer a valid assumption to hardcode the `Memory`
backend's current working directory.

This commit transitions the vfs backend to a loader style in which
individual backends are allowed to resolve relative paths as they see
fit.

To allow `require` to continue to set the `__FILE__` magic constant,
this commit adds a new API to `artichoke_core::load::LoadSources` for
resolving the absolute path name from the vfs if it exists, which allows
individual vfs backends to perform absolutizing.

The require flow as of this commit looks like this:

- `require` resolves the absolute file path for the resource it would
  like to load via the `LoadSources::resolve_source_path` method.
- `LoadSources::resolve_source_path` delegates to the `resolve_file`
  methods in the vfs backends, which should absolutize paths relative to
  their own internal working directroy.
- `require` sets the parser `Context` with the resolved path.
- `require` requires the file via the `LoadSources::require_source`
  method.
- `LoadSources::require_source` delegates to the vfs backends which
  aboslutize the path relative to their own internal working directory.

This commit bumps the version of `artichoke-core` to 0.8.0.