Improve performance and overflow safety in spinoso-array
Use `resize` and `resize_with` for `nil` filling buffers when setting
out of bounds indexes instead of `push`ing in a loop.

Add tests for `set` with an out of bounds index.

Use checked addition when given an index and length that might overflow
`usize::MAX`. Some cases are replaced with saturating adds or indexing
to the end of the buffer; some cases are replaced with capacity overflow
panics.

Bump `spinoso-array` to 0.5.2.