Fix a memory leak when sending an :initialize message to an initialized Regexp
If a `Regexp` is already initialized, there is a heap allocated object
stored in its `RBasic` data pointer. Reinitializing the `mrb_value` will
overwrite this pointer without freeing the `Box`, which results in a
leak.

In Ruby, `Regexp` is not allowed to be reinitialized. By inspecting the
`Source` flags, we can determine if the `Regexp` was a literal and raise
the appropriate exception to match MRI.