Disable default-features for spinoso-env and spinoso-math in artichoke-backend
The `Cargo.toml` in `artichoke-backend` activates optional features in
`spinoso-env` and `spinoso-math` based on its own feature flags. This
means the dependency declarations should disable default features.

This cuts down on compiled code in some feature flag configurations and
fixes a bug where dependent APIs might be exposed even if they are
intended to be disabled.