Fix file not found error when running the spec-runner
PR #1299 reworked file loading such that `.rb` suffix appending only
happens in `require`.

The `spec-runner` uses `Eval::eval_file` but supplies a non-existant
path. This PR fixes the bad path and adds a CI step to run the
spec-runner tests in CI.

This breakage made it to master where it failed the LeakSan build for
the `spec-runner`.