Remove `std` feature from `spinoso-math`
`spinoso-math` documents itself to be `no_std` and has a `std` feature,
but the crate did not declare the `#![no_std]` pragma.

Many `f64` math functions require `std` and are not available in `core`.
Despite missing support in `core`, without the crate properly declaring
itself as `no_std`, tests with `--no-default-features` incorrectly
passed since `std` is still linked in.

It does not seem likely that `f64` math will be enabled in `core` any
time soon, so make `spinoso-math` require `std`. See rust-lang/rust#63455.

This commit bumps the version of `spinoso-math` to 0.2.0 since removing
a feature is a breaking change (and did break the build of
`artichoke-backend`).