Add YAML and Summary formatters to `spec-runner` output
This commit adds an optional `--format` flag to the `spec-runner` which
allows for different output formats. In addition to the existing
`artichoke` formatter, this commit adds support for a `summary`
formatter and a `yaml` formatter.

The `summary` and `yaml` formatters are based off of the built-in
formatters in `MSpec`. These formatters output raised exceptions and
backtraces as well as summary statistics (including elapsed time as
enabled by #1318).

The `summary` and `yaml` formatters do not do any exception filtering
for known failing cases like the `artichoke` formatter does. This makes
these formatters appropriate for computing historical spec pass rates.