Rebox modified `Array` into `mrb_value` before alloc in `shift`
If an `Integer` argument is given to `Array#shift`, multiple items are
removed from the front of the `Array`'s underlying vector. These removed
elements are returned from `Array#shift` as a new `Array`. `Array#shift`
with an `Integer` argument is similar to an inverted `Vec::split_off` from
the Rust std.

The the array trampoline in `artichoke-backend` failed to repack the raw
parts of the `Array` into the receiver `mrb_value` before allocating a
new `Array` for the shifted elements.

All allocations in the mruby heap can trigger a garbage collection. If
the allocation of the new `Array` for the shifted elements triggered a
GC, as it does in the `spec-runner` with `yaml` formatter as outlined in
issue #1320, the mruby GC will attempt to mark all of the children in
the original receiver. The pointers in the receiver's `RArray *` will be
garbage, so attempting to index into the raw pointer will trigger
undefined behavior, which was expressing as a segfault.

This commit re-orders the code to ensure the receiver is repacked before
attempting to allocate a new `Array` for the result of `shift`.

The following invocation of the spec-runner succeeds as of this commit:

    ./target/debug/spec-runner --format yaml all-core-specs.toml

Fixes #1320.