Move large `String::chomp` implementation out of a generic method
Extract the `String::chomp` business logic out of the generic method
which accepts the separator as an `AsRef<[u8]>` generic. This change
makes the `String::chomp` method a generic trampoline to a chomp
implementation that uses concrete `&[u8]`. This minimizes the amount of
duplicated code if `String::chomp` is called with multiple different
types for the separator argument.