Move `String::index` and `String::rindex` logic out of generic context
Move `String::index` and `String::rindex` business logic out of the
generic entry point which accepts the needle as an `AsRef<[u8]>` generic
type.

The business logic is moved to an inner function with a concrete
`needle: &[u8]` type. This minimizes the amount of duplicated code as
these functions are generated when called with different types for the
needle.