Add GitHub Actions councurrency limits to workflows that `git push`
Occasionally when multiple PRs are merged to trunk quickly, the rustdoc
and spec-state workflows fail due to a race between generating a new
commit to push and another workflow instance completing a push.

Limit workflow concurrency to one concurrent run per branch to ensure
pushes to trunk are processed serially.

https://github.blog/changelog/2021-04-19-github-actions-limit-workflow-run-or-job-concurrency/