Add an encoding-aware character indexing API to spinoso-string
`String::get` in `spinoso-string` operates on a byte level and has
semantics most close to `String#getbyte` and `String#byteslice` in Ruby.

The various overloads of `String#[]` are encoding-aware. `"💎🦀"[1]`
returns the second UTF-8 character in the string, "🦀".

Add `get_char`, which is an encoding-aware version of `get` with a
single usize `SliceIndex`.