Improve handling of multi-byte UTF-8 sequences in `get_char_slice`
The fast path to consume ASCII bytes when computing the byte slice
offset of the selected substring is only valid if the string is ASCII up
to the end of the requested char subslice of the string.

If the first non-ascii character occurs before the requested offset,
consume up to the first non-ascii char.

If the first non-ascii char occurs somewhere else, abort the fast path
and consume no bytes.