Add an encoding-aware substring indexing API to spinoso-string
`String::get` in `spinoso-string` operates on a byte level and has
semantics most close to `String#getbyte` and `String#byteslice` in Ruby.

The various overloads of `String#[]` are encoding-aware. `"💎🦀"[0, 2]`
returns a two UTF-8 character-long substring, "💎🦀".

Add `get_char_slice`, which is an encoding-aware version of `get` with a
`Range<usize>` `SliceIndex`.