Improve readability of UTF-8 slow path in `get_char_slice`
Refactor the slow path in the UTF-8 encoding case to be a decode loop
over a big match statement on the results of `bstr::decode_utf8`. This
section is similar to that used in `get_char` as introduced in the prior
commit.

Add explainer comments to some of the fast paths that attempt to short
circuit the slow char-at-a-time decode loop.