Remove potential for stack overflow when booting interpreter
Prior to this commit, all core and stdlib modules in `extn` that loaded
Ruby sources into the interpreter did so with `include_bytes!` macro
invocations. These calls were done inline in the various `init`
functions.

`include_bytes!` returns a fixed-sized array, so all of the Ruby source
contents were on the stack. For modules like `JSON` and `URI`, this is a
lot of data on the stack.

Artichoke intends to be embeddable and makes no assumptions about how
deep the stack is in the application when the interpreter is booted.

Instead of calling `include_bytes!` inline, move the `include_bytes!`
calls to store the byte contents in a `static` binding. `static`s are
stored at a fixed location in memory and the program binary. These
statics are coerced to `&[u8]` which means when they are used in `init`,
the only stack space used is for a fat pointer.

This commit also removes largely useless calls to `trace!` and `debug!`
macros when bootstrapping the interpreter with individual classes and
modules.

`numeric::init` is moved to `numeric::mruby::init`.